Conflicts:
	scilab/configure
	scilab/modules/ast/Makefile.am
	scilab/modules/ast/Makefile.in
	scilab/modules/ast/ast.vcxproj
	scilab/modules/ast/ast.vcxproj.filters
	scilab/modules/core/includes/version.h.in
	scilab/modules/core/includes/version.h.vc
	scilab/modules/functions/Makefile.in
	scilab/modules/symbolic/Makefile.in

+++ scilab/modules/ast/includes/exps/smallintselectexp.hxx
@@ -62,8 +62,8 @@ public:
 
     inline Exp * getExp(const int64_t key) const
     {
-        const unsigned int size = table.size();
-        for (unsigned int i = 0; i < size; ++i)
+        const int size = static_cast<int>(table.size());
+        for (int i = 0; i < size; ++i)
         {
             if (table[i] == key)
             {
